// src/app/generated/models/contractor-contractor-jobs-data.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import {ContractorJob} from './contractor-job';
import {FinancialPartialData} from './financial-partial-data';

export interface ContractorContractorJobsData {
    averageFinancialData?: FinancialPartialData;
    contractorJobs?: Array<ContractorJob>;
    financialData?: FinancialPartialData;
}